ObjectAnimation: Split Object animation code into new file
