+++ game/src/game/CharacterMatrix.h
@@ -30,3 +30,4 @@ class Object;
 bool    chr_matrix_valid( const Object * pchr );
 egolib_rv chr_update_matrix( Object * pchr, bool update_size );
 bool set_weapongrip( const CHR_REF iitem, const CHR_REF iholder, uint16_t vrt_off );
+bool chr_getMatUp(Object *pchr, fvec3_t& up);
